fix(gm): address review on chat-command PR (#517)
Resolve CodeRabbit + Copilot review findings:

- Split the 1108-line cell/gm_command.rs (over the 700-line hard cap) into a
  gm_command/ module directory: mod.rs (dispatch + re-exports), feedback,
  spawn, movement, combat_cmd, inventory, query. Public surface unchanged
  (handle_gm_command + send_gm_feedback re-exported); pure mechanical move,
  no behavior change. All 15 handler tests co-located and retained.
- Log on cell-send failure in the base GM-command interception (was a silent
  `let _ = tx.send(...)`): a dropped command/feedback now emits a warn with
  player + command context, per the negative-logging convention.
- Fix the /goto <player> docs: players resolve by numeric entity id; name
  matching only works for named entities (NPCs), not player characters.
  Added a clarifying comment at the name-resolution branch.
- Strip volatile PR/issue numbers from commands.md (kept the stable CAT-N-04
  finding id) per copilot-instructions.
